% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{as.character.mfactor}
\alias{as.character.mfactor}
\title{Transforms an mfactor to character depending to the underlying representation needed}
\usage{
\method{as.character}{mfactor}(x, ...)
}
\arguments{
\item{x}{mfactor}
}
\description{
Transforms an mfactor to character depending to the underlying representation needed
}
