% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{as.data.frame.mfactor}
\alias{as.data.frame.mfactor}
\title{Transforms an mfactor to data frame}
\usage{
\method{as.data.frame}{mfactor}(x, ...)
}
\arguments{
\item{x}{mfactor}

\item{...}{Arguments passed on to \code{as.data.frame.vector}
\describe{
  \item{}{}
}}
}
\description{
Transforms an mfactor to data frame
}
