% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as_factor_mfactor.R
\name{as.factor.mfactor}
\alias{as.factor.mfactor}
\title{Coersion to ordinary factor vectors from multifactor vectors}
\usage{
\method{as.factor}{mfactor}(x, collapse, sep = getOption("mfactor.sep", ";"),
  ..., none = if (any((!is.na(unclass(x))) & unclass(x) == 0))
  getOption("mfactor.none", "<None>") else NA)
}
\arguments{
\item{x}{a multi-factor to be coerced to a factor}

\item{collapse}{Either (A) an *aggregating function*, which returns a
single value when supplied a integer vector input (such as `min`,`max`, and
`median`, etc).  in order to represent the mutiply valued elment in the the
resulting vactor, or (B) a string which should replace multiply valued
elements in the resulting factor, or (C) `NA`.  Note that only one of the
arguments `sep` and `collapse` should be specified.}

\item{sep}{A character string used to separate the distinct values that an entry may
defaults to ';'}

\item{...}{Additional parameters \code{...} are is not supported}

\item{none}{A character representation for an element in which none of the levels are
taken.  Defaults to \code{'getOption('mfactor.none','<None>')'}.  NOTE that \code{'NONE'}
is not the same as NA, the latter representing no information whereas the former indicates
the knowlege that none of levels were indicated.}
}
\description{
Coersion to ordinary factor vectors from multifactor vectors.
}
\seealso{
Other Coercion-from-mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}},
  \code{\link{ordered.mfactor}};
  \code{\link{mfactor.mfactor}}; \code{\link{mfactor}}
}

