% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{as.integer.mfactor}
\alias{as.integer.mfactor}
\title{Transforms an mfactor to the internal integer representation}
\usage{
\method{as.integer}{mfactor}(x, ...)
}
\arguments{
\item{x}{mfactor}
}
\description{
Transforms an mfactor to the internal integer representation
}
