% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{as.mfactor.default}
\alias{as.mfactor.default}
\title{Transformation to mfactor}
\usage{
\method{as.mfactor}{default}(x, ...)
}
\arguments{
\item{x}{value}

\item{...}{Arguments passed on to \code{as.mfactor}
\describe{
  \item{x}{values to transform}
}}
}
\description{
Transformation to mfactor
}
