% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{as.mfactor}
\alias{as.mfactor}
\title{Generic function to overload for casting to mfactor}
\usage{
as.mfactor(x, ...)
}
\arguments{
\item{x}{values to transform}

\item{...}{Arguments passed on to \code{as}
\describe{
  \item{}{}
}}
}
\description{
Generic function to overload for casting to mfactor
}
