% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compatibilityCheck.R
\name{compatibilityCheck}
\alias{compatibilityCheck}
\title{Test that two (x and y) data.frame's are compatibile for binding via \code{rbind(x,y)}.}
\usage{
compatibilityCheck(x, y, verbose = TRUE, .stop = FALSE)
}
\arguments{
\item{x,y}{\code{data.frame} objects}

\item{verbose}{If TRUE, and errors are present, the table of errors is printed to the console.}

\item{.stop}{If TRUE, and errors are present, execution is stopped via `stop()`.}
}
\description{
Identifies x variety of incompatibilities between two \code{data.frame} objects
that could cause errors, or missing data when combined via \code{rbind(x,y)}
and raises an error if combining the two data.frames would result in missing
data or other unexpected outcomes.
}
\details{
Note that setting \code{'options(strict.Rbind=TRUE)'} will cause this compatibilty check
to be run whenever rbind(x,y) is called with two \code{data.frame} objects.
}

