% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as_factor.R, R/factor.R, R/ordered.R
\name{as.factor}
\alias{as.factor}
\alias{as.ordered}
\alias{factor}
\alias{ordered}
\title{S3 generics for \code{\link[base]{factor}},
\code{\link[base]{ordered}} ,
\code{\link{as.factor}}, and \code{\link{as.ordered}}}
\usage{
as.factor(x, ...)

factor(x, ...)

ordered(x, ...)

as.ordered(x, ...)
}
\arguments{
\item{x}{a vector of data, usually taking a small number of distinct values.}

\item{...}{additional parameters to be passed to specific methods of \code{\link[base]{factor}},
\code{\link[base]{ordered}} or appropriate methods such as
\code{\link{factor.mfactor}} and \code{\link{ordered.mfactor}}}
}
\description{
S3 generics for \code{\link[base]{factor}},
\code{\link[base]{ordered}} ,
\code{\link{as.factor}}, and \code{\link{as.ordered}}
}

