% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{format.mfactor}
\alias{format.mfactor}
\title{Formats en element of an mfactor}
\usage{
\method{format}{mfactor}(x, ...)
}
\arguments{
\item{x}{mfactor element}
}
\description{
Formats en element of an mfactor
}
