% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is_mfactor.R
\name{is.mfactor}
\alias{is.mfactor}
\title{Test if an object is a multifactor}
\usage{
is.mfactor(x)
}
\arguments{
\item{x}{An object to be tested}
}
\description{
Test if an object is a multifactor.
}
\examples{
is.mfactor(1)
is.mfactor('a')
is.mfactor(mfactor('a'))
}

