% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as_matrix_mfactor.R, R/mfactor_matrix.R
\name{as.matrix.mfactor}
\alias{as.matrix.mfactor}
\alias{mfactor.matrix}
\title{Conversion between indicator matricies and mfactor variables}
\usage{
\method{as.matrix}{mfactor}(x, ...)

\method{mfactor}{matrix}(x, levels = dimnames(x)[[2]], ...)
}
\arguments{
\item{x}{A matrix of mode logical to be coerced to a vector with class multi-factor
or a multifactor to be coerced to matrix.}

\item{...}{Additional arguments, which affect the coersion of an x to
class mfactor}

\item{levels}{a vector of levls with length equal to \code{ncol(x)}
to be the levels of the resulting mfactor. (default = dimnames(x)[[2]])}
}
\description{
Conversion between indicator matricies and mfactor variables.
}
\examples{
(x = matrix(
   c(TRUE ,TRUE ,TRUE ,FALSE,
     FALSE,FALSE,TRUE ,FALSE,
     TRUE ,TRUE ,FALSE,FALSE),
   nrow=3,
   dimnames = list(1:3,letters[1:4])))
mfactor(x)

(y = mfactor(c('1,2,3','4,5',6,7,8),split = ',',levels = 1:10))
as.matrix(y)
}
\seealso{
Other Coercion-from-mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.factor.mfactor}};
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}},
  \code{\link{ordered.mfactor}};
  \code{\link{mfactor.mfactor}}; \code{\link{mfactor}}

Other Coercion-to-mfactor: \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}}; \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}},
  \code{\link{ordered.mfactor}};
  \code{\link{mfactor.character}};
  \code{\link{mfactor.mfactor}}; \code{\link{mfactor}}
}

