% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mfactor_character.R
\name{mfactor.character}
\alias{mfactor.character}
\title{Multi-level Factors}
\usage{
\method{mfactor}{character}(x = character(), split, fixed = FALSE,
  perl = FALSE, useBytes = FALSE, ...)
}
\arguments{
\item{x}{An object to be coerced into a multi-factor}

\item{split}{
    character vector (or object which can be coerced to such)
    containing \link{regular expression}(s) (unless \code{fixed = TRUE})
    to use for splitting.  If empty matches occur, in particular if
    \code{split} has length 0, \code{x} is split into single characters.
    If \code{split} has length greater than 1, it is re-cycled along
    \code{x}.
  }

\item{fixed}{
    logical.  If \code{TRUE} match \code{split} exactly, otherwise
    use regular expressions.  Has priority over \code{perl}.
  }

\item{perl}{logical.  Should perl-compatible regexps be used?}

\item{useBytes}{logical.  If \code{TRUE} the matching is done
    byte-by-byte rather than character-by-character, and inputs with
    marked encodings are not converted.  This is forced (with a warning)
    if any input is found which is marked as \code{"bytes"}.}

\item{...}{additional arguments, depending on the value of x}
}
\description{
The function \code{mfactor} is used to create a factor like vector in which
individual entries may take zero or more values from the unique levels of x
(\code{levels(x)}).
}
\details{
mfactors are particularly useful for transforming delimited text vectors
(e.g. ) and converting it to a more useable format such as a \code{matrix}
or a set of indicator variables.  (See examples)
}
\examples{
(x = c("a,b,c","c","a,b"))
mfactor(x,levels=letters[1:4],split = ',')

(y = mfactor(c('1,2,3','4,5',6,7,8),split = ',',levels = 1:10))
as.character(y,sep = ',')
}
\seealso{
Other mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.factor.mfactor}},
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}};
  \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{compatibilityCheck}}; \code{\link{mfactor}},
  \code{\link{mfactor.mfactor}}
}

