% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mfactor_mfactor.R
\name{mfactor.mfactor}
\alias{mfactor.mfactor}
\title{Recasting of mfactor objects}
\usage{
\method{mfactor}{mfactor}(x, levels = attr(x, "levels"), labels = levels,
  ordered = inherits(x, "ord_mfactor"), ...)
}
\arguments{
\item{x}{a list where each element contains zero or more objects from
a unique set of levels.}

\item{levels}{an optional vector of the values (as character strings)
that x might have taken. The default is the unique set of values taken by
as.character(x), sorted into increasing order of x. Note that this set can
be specified as smaller than sort(unique(x)).}

\item{labels}{\strong{Either} an optional character vector of labels for the
levels (in the same order as 'levels' after removing those in
'exclude'), \strong{or} a character string of length 1.}

\item{ordered}{logical flag to determine if the levels should be regarded as
ordered (in the order given).}

\item{...}{Additional arguments, which affect the coersion of an x to
class mfactor}
}
\description{
Recasting of mfactor objects.
}
\examples{
(x = mfactor(c('a','b','c')))

# reverse the order of the levels:
mfactor(x,levels = c('c','b','a'))
}
\seealso{
Other Coercion-from-mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.factor.mfactor}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}},
  \code{\link{ordered.mfactor}}; \code{\link{mfactor}}

Other Coercion-to-mfactor: \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}},
  \code{\link{ordered.mfactor}};
  \code{\link{mfactor.character}}; \code{\link{mfactor}}
}

