% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{mfactor}
\alias{mfactor}
\title{mfactor generation}
\usage{
mfactor(x = character(), levels, labels, exclude = NA,
  ordered = is.ordered(x), nmax = NA, representation = NULL)
}
\arguments{
\item{x}{values to transform}

\item{levels}{vector defining the values x can take}

\item{labels}{dataframe representing labels for this mfactor}

\item{exclude}{values to remove from x}

\item{ordered}{defines if levels have to be ordered}

\item{nmax}{max number of levels}

\item{representation}{name of the initial representation of this mfactor}
}
\description{
mfactor generation
}
