% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{mlevels<-}
\alias{mlevels<-}
\title{Overload of the <- operator in order to define mfactor levels}
\usage{
mlevels(x) <- value
}
\arguments{
\item{x}{mfactor}

\item{value}{labels}
}
\description{
Overload of the <- operator in order to define mfactor levels
}
