% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/none.R
\docType{data}
\name{NONE}
\alias{NONE}
\alias{is.none}
\alias{is.none.default}
\alias{is.none.mfactor}
\alias{is.none.ord_mfactor}
\alias{is.none<-}
\alias{is.none<-.default}
\alias{is.none<-.mfactor}
\title{The None object}
\format{\preformatted{Class 'mfactor'  atomic [1:1] 0
  ..- attr(*, "levels")= chr(0) 
  ..- attr(*, "mlevels")= chr(0) 
}}
\usage{
NONE

is.none(x)

\\method{is.none}{default}(x)

\\method{is.none}{ord_mfactor}(x)

\\method{is.none}{mfactor}(x)

is.none(x) <- value

\\method{is.none}{default}(x) <- value

\\method{is.none}{mfactor}(x) <- value
}
\arguments{
\item{x}{an \R object to be tested: the default method for
    \code{is.na} handles atomic vectors, lists and pairlists: that for
    \code{anyNA} also handles \code{NULL}.}

\item{value}{a suitable index vector for use with \code{x}.}
}
\description{
The None object in the mfactor package is used to indicate
that none of the elements are selected, which is distinct
from 'NA' which indicates that is is unknown whether or not
any of the elements are selected.
}
\keyword{datasets}

