% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{representation<-}
\alias{representation<-}
\title{Overload of the <- operator in order to define mfactor representation}
\usage{
representation(x) <- value
}
\arguments{
\item{x}{mfactor}

\item{value}{representation name}
}
\description{
Overload of the <- operator in order to define mfactor representation
}
