% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/special_median.R
\name{lower.median}
\alias{lower.median}
\alias{median.random}
\alias{upper.median}
\title{Special median functions.}
\usage{
lower.median(x, na.rm = FALSE)

upper.median(x, na.rm = FALSE)

\method{median}{random}(x, na.rm = FALSE)
}
\arguments{
\item{a}{numeric vector containing the values whose (special) median is to be computed}
}
\description{
Special median functions, which return the upper(lower) of the two middle values
when lenght(x) is even
}
\examples{
x = c(1,2,3,3,3,4,5,6,7,8)

lower.median(x) # returns 3

upper.median(x) # returns 4

# 3 times more likely to return 3 than 4 b/c of the multiplicity of the value 3 in x
median.random(x)
}

