% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{[[.mfactor}
\alias{[[.mfactor}
\title{Keeps class and attributes when accessing subsets of the value}
\usage{
\method{[[}{mfactor}(x, ...)
}
\arguments{
\item{x}{mfactor}

\item{...}{Arguments passed on to \code{`[[`}
\describe{
  \item{}{}
}}
}
\description{
Keeps class and attributes when accessing subsets of the value
}
