% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{unique.mfactor}
\alias{unique.mfactor}
\title{Overloads the unique capacity for mfactor}
\usage{
\method{unique}{mfactor}(x, ...)
}
\arguments{
\item{x}{mfactor}

\item{...}{Arguments passed on to \code{unique.default}
\describe{
  \item{}{}
}}
}
\description{
Overloads the unique capacity for mfactor
}
