% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor-datatype.r
\name{unlistMfactor}
\alias{unlistMfactor}
\title{Unlist for mfactor elements}
\usage{
unlistMfactor(x, ...)
}
\arguments{
\item{x}{mfactor}

\item{...}{Arguments passed on to \code{unlist}
\describe{
  \item{}{}
}}
}
\description{
Unlist for mfactor elements
}
