% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_table.R
\name{write.table}
\alias{write.table}
\title{A wrapper for utils::write.table()}
\usage{
write.table(x, ...)
}
\description{
A wrapper for utils::write.table() which handles the
coercion of the mfactor variables to character before
writing them to file so that they are quoted appropropriately
}

