% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as_character_mfactor.R
\name{as.character.mfactor}
\alias{as.character.mfactor}
\title{Coersion of multi-factors to character vectors}
\usage{
\method{as.character}{mfactor}(x, sep = ";",
  none = getOption("mfactor.none", "<None>"))
}
\arguments{
\item{x}{a multi-factor to be coerced to a character vector}

\item{sep}{A character string used to separate the distinct values that an entry may
defaults to ';'}

\item{none}{A character representation for an element in which none of the levels are
taken.  Defaults to \code{'getOption('mfactor.none','<None>')'}.  NOTE that \code{'NONE'}
is not the same as NA, the latter representing no information whereas the former indicates
the knowlege that none of levels were indicated.}
}
\description{
Coersion of multi-factors to character vectors
}
\seealso{
Other mfactor: \code{\link{as.factor.mfactor}},
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}};
  \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{compatibilityCheck}};
  \code{\link{mfactor.character}}; \code{\link{mfactor}},
  \code{\link{mfactor.mfactor}}
}

