% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mFactor-rbind.R
\name{compatibilityCheck}
\alias{compatibilityCheck}
\title{Test that two (x and y) data.frame's are compatibile for binding via \code{rbind(x,y)}.}
\usage{
compatibilityCheck(a, b)
}
\arguments{
\item{x,y}{\code{data.frame} objects}
}
\description{
Identifies a variety of incompatibilities between two \code{data.frame} objects
that could cause errors, or missing data when combined via \code{rbind(x,y)}
and raises an error if combining the two data.frames would result in missing
data or other unexpected outcomes.
}
\details{
Note that setting \code{'options(strict.Rbind=TRUE)'} will cause this compatibilty check
to be run whenever rbind(x,y) is called with two \code{data.frame} objects.
}
\seealso{
Other mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.factor.mfactor}},
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}};
  \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{mfactor.character}}; \code{\link{mfactor}},
  \code{\link{mfactor.mfactor}}
}

