% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as_factor_mfactor.R, R/factor_mfactor.R, R/mfactor_factor.R
\name{as.factor.mfactor}
\alias{as.factor.mfactor}
\alias{factor.mfactor}
\alias{mfactor.factor}
\title{Coersion of multi-factors to oridnary factors}
\usage{
\method{as.factor}{mfactor}(x, collapse, sep = getOption("mfactor.sep", ";"),
  ..., none = if (any((!is.na(unclass(x))) & unclass(x) == 0))
  getOption("mfactor.none", "<None>") else NA)

\method{factor}{mfactor}(x, levels = attr(x, "levels"), labels = levels,
  exclude, ordered = inherits(x, "ord_mfactor"), ...)

\method{mfactor}{factor}(x, levels, labels = levels,
  exclude = getOption("mfactor.none", "<None>"), ordered = is.ordered(x),
  split, ...)
}
\arguments{
\item{x}{a multi-factor to be coerced to a factor}

\item{sep}{A character string used to separate the distinct values that an entry may
defaults to ';'}

\item{...}{additional arguments depending on wether 'split' is specified}

\item{none}{A character representation for an element in which none of the levels are
taken.  Defaults to \code{'getOption('mfactor.none','<None>')'}.  NOTE that \code{'NONE'}
is not the same as NA, the latter representing no information whereas the former indicates
the knowlege that none of levels were indicated.}

\item{labels}{\strong{Either} an optional character vector of labels for the
levels (in the same order as 'levels' after removing those in
'exclude'), \strong{or} a character string of length 1.}

\item{exclude}{A vector of values to be excluded when forming the set of
levels.  NOTE THAT unlike ordinary factors, this parameter defaults to
\code{getOption('mfactor.none','<None>')}.}

\item{ordered}{logical flag to determine if the levels should be regarded as
ordered (in the order given).}

\item{split}{if specified, causes x to be coerced by \code{\link{as.character.mfactor}}}

\item{x}{a factor to be coerced to a multi-factor}
}
\description{
Coersion of multi-factors to oridnary factors.

The function \code{mfactor} is used to create a factor like vector in which
individual entries may take zero or more values from the unique levels of x
(\code{levels(x)}).
}
\details{
Note that 'base::factor' is *not* an S3 generic, so `factor.mfactor()` must be
called directly or `setGenerics()` must be be called prior to calling
`factor()`
}
\seealso{
Other mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{compatibilityCheck}};
  \code{\link{mfactor.character}}; \code{\link{mfactor}},
  \code{\link{mfactor.mfactor}}

Other mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{compatibilityCheck}};
  \code{\link{mfactor.character}}; \code{\link{mfactor}},
  \code{\link{mfactor.mfactor}}

Other mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{compatibilityCheck}};
  \code{\link{mfactor.character}}; \code{\link{mfactor}},
  \code{\link{mfactor.mfactor}}
}

