% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factor_mfactor.R, R/mfactor_factor.R, R/ordered.R
\name{factor.mfactor}
\alias{factor.mfactor}
\alias{mfactor.factor}
\alias{ordered.mfactor}
\title{Coersion of multi-factors to oridnary factors}
\usage{
\method{factor}{mfactor}(x, levels = attr(x, "levels"), labels = levels,
  exclude, ordered = inherits(x, "ord_mfactor"), ...)

\method{mfactor}{factor}(x, levels, labels = levels,
  exclude = getOption("mfactor.none", "<None>"), ordered = is.ordered(x),
  split, ...)

\method{ordered}{mfactor}(x, ...)
}
\arguments{
\item{x}{a multi-factor to be coerced to a factor}

\item{...}{Additional parameters \code{...} are is not supported}
}
\description{
Coersion of multi-factors to oridnary factors.
}
\details{
Note that 'base::factor' is *not* an S3 generic, so `factor.mfactor()` must be
called directly or `setGenerics()` must be be called prior to calling
`factor()`
}
\seealso{
Other Coercion-from-mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.factor.mfactor}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{mfactor.mfactor}}; \code{\link{mfactor}}

Other Coercion-to-mfactor: \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{mfactor.character}};
  \code{\link{mfactor.mfactor}}; \code{\link{mfactor}}
}

