% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as_list_mfactor.R, R/mfactor_list.R
\name{as.list.mfactor}
\alias{as.list.mfactor}
\alias{mfactor.list}
\title{Conversion between lists of values and mfactor variables}
\usage{
\method{as.list}{mfactor}(x, levels = FALSE, ...)

\method{mfactor}{list}(x, labels = levels, levels, ordered = FALSE, exclude,
  ...)
}
\arguments{
\item{x}{a list where each element contains zero or more objects from
a unique set of levels.}

\item{levels}{an optional vector of the values (as character strings)
that x might have taken. The default is the unique set of values taken by
as.character(x), sorted into increasing order of x. Note that this set can
be specified as smaller than sort(unique(x)).}

\item{...}{Additional arguments, which affect the coersion of an x to
class mfactor}

\item{labels}{\strong{Either} an optional character vector of labels for the
levels (in the same order as 'levels' after removing those in
'exclude'), \strong{or} a character string of length 1.}

\item{ordered}{logical flag to determine if the levels should be regarded as
ordered (in the order given).}

\item{exclude}{A vector of values to be excluded when forming the set of
levels.  NOTE THAT unlike ordinary factors, this parameter defaults to
\code{getOption('mfactor.none','<None>')}.}
}
\description{
Conversion between lists of values and mfactor variables.
}
\examples{
(x = list(c('a','b','c'),c('c'),c('a','b')))
mfactor(x,levels=letters[1:4])

(y = mfactor(c('1,2,3','4,5',6,7,8),split = ',',levels = 1:10))
as.list(y)
}
\seealso{
Other Coercion-to-mfactor: \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}},
  \code{\link{ordered.mfactor}};
  \code{\link{mfactor.character}};
  \code{\link{mfactor.mfactor}}; \code{\link{mfactor}}

Other Coercion-to-mfactor: \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}},
  \code{\link{ordered.mfactor}};
  \code{\link{mfactor.character}};
  \code{\link{mfactor.mfactor}}; \code{\link{mfactor}}
}

