% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mFactor.R
\name{mfactor}
\alias{mfactor}
\alias{mfactor.mfactor}
\title{Multi-level Factors}
\usage{
mfactor(x, ...)

\method{mfactor}{mfactor}(x, levels = attr(x, "levels"), labels = levels,
  ordered = inherits(x, "ord_mfactor"), ...)
}
\arguments{
\item{x}{An object to be coerced into a multi-factor
levels an optional vector of the values (as character strings) that
'x' might have taken.  The default is the unique set of
values taken by \code{'as.character(strsplit(x,split,...))'} if the
argument \code{'split'} is specified and \code{'as.character(x)'} otherwise,
sorted into increasing order _of 'x'_.  Note that the value of this argument
does not have to include all possible values of x.}

\item{...}{additional arguments, depending on the value of x.}

\item{labels}{\strong{Either} an optional character vector of labels for the
levels (in the same order as 'levels' after removing those in
'exclude'), \strong{or} a character string of length 1.}

\item{ordered}{logical flag to determine if the levels should be regarded as
ordered (in the order given).}

\item{exclude}{A vector of values to be excluded when forming the set of
levels.  NOTE THAT unlike ordinary factors, this parameter defaults to
\code{getOption('mfactor.none','<None>')}.}
}
\description{
The function \code{mfactor} is used to create a factor like vector in which
individual entries may take zero or more values from the unique levels of x
(\code{levels(x)}).
}
\seealso{
Other mfactor: \code{\link{as.character.mfactor}};
  \code{\link{as.factor.mfactor}},
  \code{\link{factor.mfactor}},
  \code{\link{mfactor.factor}};
  \code{\link{as.list.mfactor}},
  \code{\link{mfactor.list}};
  \code{\link{as.matrix.mfactor}},
  \code{\link{mfactor.matrix}};
  \code{\link{compatibilityCheck}};
  \code{\link{mfactor.character}}
}

