% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/special_median.R
\name{lower.median}
\alias{lower.median}
\alias{median.random}
\alias{upper.median}
\title{Special median functions.}
\usage{
lower.median(x, na.rm = FALSE)

upper.median(x, na.rm = FALSE)

median.random(x,na.rm=FALSE)
}
\arguments{
\item{x}{an object for which a method has been defined, or a numeric vector containing the values whose median is to be computed.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
When `length(x)`, these special median functions return the upper, lower, or a random
selection from the middle two values of x.  When `length(x)` is odd, each function
is performes identcialy to `base::median`.

A randomly selected median.
}
\examples{
x = c(1,2,3,3,3,4,5,6,7,8)

lower.median(x) # returns 3

upper.median(x) # returns 4

# 3 times more likely to return 3 than 4 b/c of the multiplicity of the value 3 in x
median.random(x)
}

